+++ NEWS.md
@@ -1,6 +1,7 @@
 # random.cdisc.data 0.3.9.9000
 * Fix bug to make internal helper `rel_var` work with `tibbles`.
 * Added new random dataset generator: Hys Law Analysis Dataset (`radhy`) .
+* Updated the sampling of `TRTSDTM` in `radsl` to be more efficient.
 
 # random.cdisc.data 0.3.9
 * Fix warnings of binding factors to characters.